+++ src/child-process.ts
@@ -1,0 +1,4 @@
+import * as util from 'util'
+import * as childProcess from 'child_process'
+
+export const exec = util.promisify(childProcess.exec)