+++ src/git.ts
@@ -1,9 +1,5 @@
-// tslint:disable:prettier
-import * as util from "util"
-import * as childProcess from "child_process"
-import * as stringMap from "@ts-common/string-map"
-
-const exec = util.promisify(childProcess.exec)
+import * as childProcess from './child-process'
+import * as stringMap from '@ts-common/string-map'
 
 export type ExecResult = {
   readonly stdout: string
@@ -13,19 +9,19 @@ export type ExecResult = {
 export type GenericCommand = stringMap.StringMap<readonly string[]>
 
 export type Command =
-  { readonly config: readonly ["user.email" | "user.name", string] } |
+  { readonly config: readonly ['user.email' | 'user.name', string] } |
   { readonly init: readonly [] } |
   { readonly add: readonly [string] } |
-  { readonly commit: readonly ["-m", string, "--no-gpg-sign"] } |
-  { readonly checkout: readonly ["-b", string]|readonly [string] } |
-  { readonly branch: readonly [string]|readonly [string, string] } |
-  { readonly remote: readonly ["add", string, string] } |
+  { readonly commit: readonly ['-m', string, '--no-gpg-sign'] } |
+  { readonly checkout: readonly ['-b', string] | readonly [string] } |
+  { readonly branch: readonly [string] | readonly [string, string] } |
+  { readonly remote: readonly ['add', string, string] } |
   { readonly clone: readonly [string, string] } |
-  { readonly diff: readonly ["--name-status" | "--name-only", "--no-renames", string, string] }
+  { readonly diff: readonly ['--name-status' | '--name-only', '--no-renames', string, string] }
 
 export const repository = (repositoryPath: string) =>
   async (command: Command) => {
     const g: GenericCommand = command
     const [cmd, args] = stringMap.entries(g).toArray()[0]
-    return await exec(`git ${cmd} ${args.join(" ")}`, { cwd: repositoryPath })
+    return childProcess.exec(`git ${cmd} ${args.join(' ')}`, { cwd: repositoryPath })
   }