+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.4.2
+
+- restore test coverage
+- use Jest instead of Mocha and Nyc
+- use `tslint` and `tslint-config-standard` original rules instead of `prettier`
+
 ## 0.4.1
 
 - `git` namespace.